requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
APScheduler==3.10.4
python-dateutil==2.8.2
requests==2.31.0
twilio==9.10.1
openai==1.6.1
gunicorn==21.2.0
python-dotenv==1.0.0
